Removing `wait_for_response`
This is due to skipping creating a new coroutine to wait for te answer. 
This might introduce a race condition but it's worth since the side 
effects are not critical. 

This solves a big problem which was starting a new `rpc_request` from 
the `ExecutorThread`. Before it couldn't find an event loop if called 
from `on_event`, for example but now that is just a blocking piece of 
code, it does work much more like expected.